requirements.txt
asgiref==3.3.2
Django==3.1.7
# This is the most recent version that can be used with Python 3.6
# As per https://docs.scipy.org/doc/scipy/reference/toolchain.html
# the version history is here https://pypi.org/project/numpy/#history
numpy==1.19.5
# This is the most recent version that can be used with Python 3.6
# As per https://pypi.org/project/pandas/1.1.5/ (see the Programming Language
# section on the left) the version history is here
# https://pypi.org/project/pandas/#history
pandas==1.1.5
psycopg2==2.9.1
pymongo==3.12.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.3.0
azure-common==1.1.27
azure-core==1.17.0
azure-identity==1.6.1
azure-keyvault-secrets==4.3.0